Allow services user to boot vms with ports on the public nets
Starting from juno services users are not granted with the admin
role anymore but with a new `service` role.
Since nova policy allows only vms booted by admin users to attach
ports on the public networks, we need to modify the policy and
allow users with the service to do that too.

Change-Id: I21506b9dadc689dfec5e5fe776746503cecc4c97
Signed-off-by: Rosario Di Somma <rosario.disomma@dreamhost.com>