DreamHost: Fix the floating ip allocation issue
DHC-2977

A recent upstream change is not compatible with the way we are
patching neutron to work with akanda and is broking floating ip
allocation(commit 22ef9e99b1b4fa28ceb3f2b77d67a1e19494b6b3)
The change is using the device_id to filter ports passing as value
the ID of the neutron router. For ports attached to routersm akanda
set that value to be the ID ot the akanda router making that query
always an empty list.

Change-Id: I2d3008b11722037b32fc30a2bf5d1cdf0886354d
Signed-off-by: Rosario Di Somma <rosario.disomma@dreamhost.com>